organize scripts folder
