+++ scripts/utilities/calcOutcomeTokenCount.js
